refactor(workspace): route becomes a package; command-name sets live in its constants
route/ now holds types (the Consumer enum), constants (the command-name
policy sets), and the route function, mirrored in both languages. The
scattered name sets move next to NAMESPACE_COMMANDS: JOB_BUILTINS from
the executor and NO_FOLLOW_COMMANDS from the links builtin. Import
path mirage.workspace.route is unchanged.
