+++ python/mirage/workspace/executor/builtins/__init__.py
@@ -15,8 +15,7 @@
 from mirage.workspace.executor.builtins.condition import handle_test
 from mirage.workspace.executor.builtins.dirs import handle_cd
 from mirage.workspace.executor.builtins.history import handle_history
-from mirage.workspace.executor.builtins.links import (NO_FOLLOW_COMMANDS,
-                                                      follow_paths, handle_ln,
+from mirage.workspace.executor.builtins.links import (follow_paths, handle_ln,
                                                       handle_readlink,
                                                       link_flags, prepare_mv,
                                                       strip_link_operands)
@@ -57,7 +56,6 @@ __all__ = [
     'handle_local',
     'handle_readlink',
     'link_flags',
-    'NO_FOLLOW_COMMANDS',
     'follow_paths',
     'prepare_mv',
     'strip_link_operands',
