+++ python/mirage/workspace/route/__init__.py
@@ -1,0 +1,29 @@
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+from mirage.workspace.route.constants import (JOB_BUILTINS, NAMESPACE_COMMANDS,
+                                              NO_FOLLOW_COMMANDS,
+                                              UNSUPPORTED_BUILTINS)
+from mirage.workspace.route.route import route
+from mirage.workspace.route.types import SHELL_CONSUMERS, Consumer
+
+__all__ = [
+    "Consumer",
+    "JOB_BUILTINS",
+    "NAMESPACE_COMMANDS",
+    "NO_FOLLOW_COMMANDS",
+    "SHELL_CONSUMERS",
+    "UNSUPPORTED_BUILTINS",
+    "route",
+]
