+++ python/mirage/workspace/route/route.py
@@ -12,51 +12,10 @@
 # limitations under the License.
 # ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
-from enum import Enum, auto
-
-from mirage.shell.types import ShellBuiltin
 from mirage.workspace.mount import MountRegistry
+from mirage.workspace.route.constants import NAMESPACE_COMMANDS, SHELL_NAMES
+from mirage.workspace.route.types import Consumer
 from mirage.workspace.session import Session
-
-# Bash builtins the parser accepts but the executor cannot honor; they
-# still route to the shell layer so the error names a capability gap.
-UNSUPPORTED_BUILTINS = frozenset({
-    "bg",
-    "disown",
-    "exec",
-    "complete",
-    "compgen",
-    "ulimit",
-})
-
-NAMESPACE_COMMANDS = frozenset({"ln", "readlink"})
-
-_SHELL_NAMES = frozenset(str(b) for b in ShellBuiltin) | UNSUPPORTED_BUILTINS
-
-
-class Consumer(Enum):
-    """The layer that consumes a command: a command belongs to the layer
-    whose state it mutates.
-
-    The verdict drives both the dispatch branch and the word policy:
-    SESSION / NAMESPACE / FUNCTION words are shell-resolved (bash
-    contract: programs receive matches, never patterns); MOUNT words
-    keep glob patterns intact for backend pushdown; UNKNOWN words are
-    never resolved (the command fails, backend I/O for it is waste).
-    """
-
-    SESSION = auto()
-    NAMESPACE = auto()
-    FUNCTION = auto()
-    MOUNT = auto()
-    UNKNOWN = auto()
-
-
-SHELL_CONSUMERS = frozenset({
-    Consumer.SESSION,
-    Consumer.NAMESPACE,
-    Consumer.FUNCTION,
-})
 
 
 def route(name: str, session: Session, registry: MountRegistry) -> Consumer:
@@ -71,7 +30,7 @@ def route(name: str, session: Session, registry: MountRegistry) -> Consumer:
         session (Session): shell session (function table).
         registry (MountRegistry): mount registry (command registration).
     """
-    if name in _SHELL_NAMES:
+    if name in SHELL_NAMES:
         return Consumer.SESSION
     if name in NAMESPACE_COMMANDS:
         return Consumer.NAMESPACE
