+++ typescript/packages/core/src/workspace/executor/builtins/index.ts
@@ -16,7 +16,6 @@ export type { ExecuteStringFn, Result } from './scope.ts'
 export { scopePath, toScope } from './scope.ts'
 export { handleCd } from './dirs.ts'
 export {
-  NO_FOLLOW_COMMANDS,
   followPaths,
   handleLn,
   handleReadlink,
